fix: failing tests on RemoteConsoleIT, RemoteGremlinIT, JsFunctionsTest (#4346)
(cherry picked from commit a68f6637e6c91b1151df5250030b78f6fe7d5839)

+++ gremlin/src/test/java/com/arcadedb/server/gremlin/RemoteGremlinIT.java
@@ -54,7 +54,7 @@ class RemoteGremlinIT extends AbstractGremlinServerIT {
         }
 
         try (final ResultSet list = graph.gremlin("g.V().hasLabel(\"inputstructure\").count()").execute()) {
-          assertThat(list.nextIfAvailable().<Integer>getProperty("result")).isEqualTo(1_000);
+          assertThat(list.nextIfAvailable().<Long>getProperty("result")).isEqualTo(1_000);
         }
 
         graph.tx().commit();
